Codes/0. DeepLab code sample/requirements.txt
numpy
scikit-learn
scipy
pillow
